created register page and its functionality. Add automatically logout if user is inactive feature.

+++ frontend/src/pages/SignupPage.tsx
@@ -1,59 +1,71 @@
 import { useContext, useEffect, useState } from 'react'
-import { Link, useLocation, useNavigate } from 'react-router-dom'
+import { useLocation, useNavigate } from 'react-router-dom'
 import { Store } from '../storeData'
+import { useSignupMutation } from '../hooks/userHook'
+import { getError } from '../utilities'
 import { ApiError } from '../types/ApiError'
 import { toast } from 'react-toastify'
-import { getError } from '../utilities'
+import 'react-toastify/dist/ReactToastify.css'
 import { Button, Container, Form } from 'react-bootstrap'
 import { Helmet } from 'react-helmet-async'
-import '../index.css'
-import LoadingBox from '../components/LoadingBox'
-import { useSigninMutation } from '../hooks/userHook'
+import { Link } from 'react-router-dom'
 
-export default function SigninPage() {
+export default function SignupPage() {
   const navigate = useNavigate()
   const { search } = useLocation()
   const redirectInUrl = new URLSearchParams(search).get('redirect')
   const redirect = redirectInUrl ? redirectInUrl : '/'
 
+  const [name, setName] = useState('')
   const [email, setEmail] = useState('')
   const [password, setPassword] = useState('')
+  const [confirmPassword, setConfirmPassword] = useState('')
 
   const { state, dispatch } = useContext(Store)
   const { userInfo } = state
 
-  const { mutateAsync: signin, isPending } = useSigninMutation()
-  const submitLoginHandler = async (e: React.SyntheticEvent) => {
+  useEffect(() => {
+    if (userInfo) {
+      navigate(redirect)
+    }
+  }, [redirect, navigate, userInfo])
+
+  const { mutateAsync: signup } = useSignupMutation()
+
+  const submitRegisterHandler = async (e: React.SyntheticEvent) => {
     e.preventDefault()
+    if (password !== confirmPassword) {
+      toast.error('Password do not match.')
+      return
+    }
     try {
-      const data = await signin({
+      const data = await signup({
+        name,
         email,
         password,
       })
       dispatch({ type: 'User_Signin', payload: data })
       localStorage.setItem('userInfo', JSON.stringify(data))
-      navigate(redirect)
+      navigate(redirect || '/')
     } catch (err) {
       toast.error(getError(err as ApiError))
     }
   }
-  useEffect(() => {
-    if (userInfo) {
-      navigate(redirect)
-    }
-  }, [navigate, redirect, userInfo])
-
   return (
     <Container className="small-container">
       <Helmet>
-        <title>Sign In</title>
+        <title>Sign Up</title>
       </Helmet>
-      <h1 className="my-3">Sign In</h1>
-      <Form onSubmit={submitLoginHandler}>
+      <h1 className="my-3"> Sign Up</h1>
+      <Form onSubmit={submitRegisterHandler}>
+        <Form.Group className="mb-3" controlId="name">
+          <Form.Label>Name</Form.Label>
+          <Form.Control onChange={(e) => setName(e.target.value)} required />
+        </Form.Group>
+
         <Form.Group className="mb-3" controlId="email">
           <Form.Label>Email</Form.Label>
           <Form.Control
-            className="custom-border"
             type="email"
             required
             onChange={(e) => setEmail(e.target.value)}
@@ -62,23 +74,25 @@ export default function SigninPage() {
         <Form.Group className="mb-3" controlId="password">
           <Form.Label>Password</Form.Label>
           <Form.Control
-            className="custom-border"
             type="password"
             required
             onChange={(e) => setPassword(e.target.value)}
           />
         </Form.Group>
+        <Form.Group className="mb-3" controlId="confirm-password">
+          <Form.Label>Confirm Password</Form.Label>
+          <Form.Control
+            type="password"
+            required
+            onChange={(e) => setConfirmPassword(e.target.value)}
+          />
+        </Form.Group>
         <div className="mb-3">
-          <Button disabled={isPending} type="submit">
-            Login
-          </Button>
-          {isPending && <LoadingBox />}
+          <Button type="submit">Register</Button>
         </div>
-
         <div className="mb-3">
-          <Link to={`/signup?redirect=${redirect}`}>
-            Don't have an account yet?
-          </Link>
+          Already have an Account?{' '}
+          <Link to={`/signin?redirect=${redirect}`}> Sign-In</Link>
         </div>
       </Form>
     </Container>
